manage states form a context management component
