+++ src/components/MainHeader/Navigation.js
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 
-import AuthContext from "../../context/auth-context";
+import AuthContext from "../../store/auth-context";
 import classes from "./Navigation.module.css";
 
 const Navigation = () => {
