--- src/context/auth-context.js
@@ -1,8 +1,0 @@
-import React from "react";
-
-const AuthContext = React.createContext({
-  isLoggedIn: false,
-  onLogout: () => {},
-});
-
-export default AuthContext;
